Created splash screen with relevant buttons
Created the initially loaded screen, with the app title, photo for a logo placeholder, and the various buttons as seen on the clickable prototype. Implemented an onClick(View) method in the MainActivity file to handle click events on the buttons. Currently, no button does anything, and the design and images are all subject to change.

Signed-off-by: AmadeusSko <amadeush@live.unc.edu>