Revised view for advanced setup, and handled inputs and a checker for when the "Finish" button is pressed.
As seen during the team meeting, there were issues with the advanced setup view not properly displaying on some devices, and the screen got a bit messy when entering
values in and still having the final button menu still visible above the keyboard. As a result, I consolidated the view to include the buttons as a bottom level element,
below all of the entry EditText items, and fixed the hanging-off-the-screen issue. I also added limited functionality to the "Go Home" button, and some functionality
to the "Finish" button which will check that all necessary fields have entries and will pull the entered-in data, but the program doesn't yet save this data anywhere.
The next commit on this branch will add a way to save that data to the device and pass it on to the Experiment Runtime class.

Signed-off-by: AmadeusSko <amadeush@live.unc.edu>