Incorporated saving and loading of inputted variables across advanced setup and runtime view.
Created a set of layouts that will display to the user what the inputted / loaded variables are.
Currently the graph view still does not actually do anything, but is otherwise ready. Allows the
user to go back to variable entry without having lost progress. Allows transfer of inputted variables
to SharedPreferences and then reading them from the same place in the Experiment Runtime view.
At this point I believe the current path to be complete. Basic UI for the walking skeleton is
complete to the extent currently possible. Next actions will be to integrate the fake data generator
upon pressing the "run" button and display that on the GraphView. This or to allow for a data stream
between the app and Arduino to run and receive actual experiment and data. Finally, implementation of an
exporting feature will also need to be added.

Signed-off-by: AmadeusSko <amadeush@live.unc.edu>