Added a full set of guided setup screens per the client's direction
Added a screen for each step of the provided app script, with text entry
points and displays of example graphs, device setup, and other displays
as requested by the client. Currently, these entries do not change any
SharedPreference values; this is pending a meeting with the client team
to determine exactly what each user selection in the guided setup will
equate to in the SharedPreferences.

Signed-off-by: AmadeusSko <amadeush@live.unc.edu>