Tweaked the experimental assistant UI for consistency and functionality
Modified some of the experimental assistant screen layouts to display
properly and allow scrolling if necessary. Also ensured that user input
for requested parameters is present, and did the same for the advanced
setup. Also fixed loading of SharedPreferences across the setups and the
experiment runtime views. Next step will be to add the user informed
consent, to define default values, and add the "About" and "Credits"
pages.

Signed-off-by: AmadeusSko <amadeush@live.unc.edu>